A ton of changes, updates, deletes.
The web is now functional, in terms of the code, the integrated code works well with the html and the CSS has made it appear to function smoothly, error handling was conducted.
Errors were removed, the major difference to this, in comparison to the previous 3 days is the, functionality, it works smoothly.

It is not complete as of yet, the notes can be found in the newly created document.

The website is at 34% completion, 77% not taking in consideration of the shop.